+++ src/main/ts/api/Main.ts
@@ -1,6 +1,7 @@
 import { Linter, Rule } from 'eslint';
 import { editor } from '../configs/Editor';
 import { standard } from '../configs/Standard';
+import { noDirectEditorEvents } from '../rules/NoDirectEditorEvents';
 import { noDirectEditorOptions } from '../rules/NoDirectEditorOptions';
 import { noDirectImports } from '../rules/NoDirectImports';
 import { noEnumsInExportSpecifier } from '../rules/NoEnumsInExportSpecifier';
@@ -12,6 +13,7 @@ import { preferFun } from '../rules/PreferFun';
 import { preferMcAgar } from '../rules/PreferMcAgar';
 
 const rules: Record<string, Rule.RuleModule> = {
+  'no-direct-editor-events': noDirectEditorEvents,
   'no-direct-editor-options': noDirectEditorOptions,
   'no-direct-imports': noDirectImports,
   'no-enums-in-export-specifier': noEnumsInExportSpecifier,