+++ src/main/ts/configs/Standard.ts
@@ -159,6 +159,7 @@ export const standard: Linter.Config = {
   ],
   rules: {
     ...base.rules,
+    '@tinymce/no-direct-editor-events': 'off',
     '@tinymce/no-direct-editor-options': 'off',
     '@tinymce/no-direct-imports': 'error',
     '@tinymce/no-enums-in-export-specifier': 'error',